[CMake] Fix usage of lowercase or in condition
CMake is a strange animal and some things are case sensitive, while
others aren't. It seems that `if(<condition>)` logical operator `OR` is
case sensitive, and `or` was not cutting it. At least for CMake 3.26.4.

If some CMake version allows `or`, it should also allow `OR`, so this
change should not be a problem for those versions.

Most of the time these pieces are not actually hit, because the build
system sets `SwiftTesting_MACROS` to `NO`, so this is skipped, which
might explain why it has not been a problem.