Remove the 'triage-needed' label from New Issue templates (#1124)
This removes the `triage-needed` label from the GitHub New Issue
templates in this repository which have it.

I added this while revamping these templates in #962, but more recently
we added a `triaged` label which is works the opposite way—it's applied
after an issue has been triaged—so we don't need the older label
anymore.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.