Infer the types of function/closure arguments when captured by an exit test.
This PR adds the ability to infer the type of a parameter of a function or closure that encloses an exit test. For example, `x` here:

```swift
func f(x: Int) async {
  await #expect(processExitsWith: .failure) { [x] in
    ...
  }
}
```

This inference still fails if a parameter is shadowed by a variable with an incompatible type; we still need something like `decltype()` to solve for such cases. Still, being able to capture `@Test` function arguments with minimal ceremony is helpful:

```swift
@Test(arguments: 0 ..< 100) func f(i: Int) async {
  await #expect(exitsWith: .failure) { [i] in
    ...
  }
}
```