Use `posix_spawn_file_actions_adddup2()` to clear `FD_CLOEXEC`.
On FreeBSD, OpenBSD, Android, and Glibc ≥ 2.29, `posix_spawn_file_actions_adddup2()` automatically clears `FD_CLOEXEC` if the file descriptors passed to it are equal. Relying on this behaviour eliminates a race condition when spawning child processes.

Some older Linuxes (Amazon Linux 2 in particular) don't have this functionality, so we do a runtime check of the Glibc version.