Synthesize display names for de facto suites with raw identifiers. (#1105)
This PR ensures that suite types that don't have the `@Suite` attribute
but which _do_ have raw identifiers for names are correctly given
display names the same way those with `@Suite` would be. This PR also
ensures that we transform spaces in raw identifiers after they are
demangled by the runtime--namely, the runtime replaces ASCII spaces (as
typed by the user) with Unicode non-breaking spaces (which aren't
otherwise valid in raw identifers) in order to avoid issues with
existing uses of spaces in demangled names. We want to make sure that
identifiers as presented to the user match what the user has typed, so
we need to transform these spaces back. No changes in this area are
needed for display names derived during macro expansion because we do
the relevant work based on the source text which still has the original
ASCII spaces.

This PR also deletes the "`raw$`" hack that I put in place when
originally implementing raw identifier support as the entire toolchain
supports them now.

Resolves #1104.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.