Have `DiscoverableAsTestContent` enumeration produce `some Sequence` instead of `AnySequence`. (#1122)
This PR changes how `DiscoverableAsTestContent` enumeration works so
that we can return `some Sequence` instead of `AnySequence`. We do so by
removing the `~Copyable` constraint on the protocol, which subsequently
causes the compiler to get confused and crash trying to represent `some
Sequence<TestContentRecord<T>>` where `T: DiscoverableAsTestContent`.

The only supported/allowed consumers of the `DiscoverableAsTestContent`
protocol are Swift Testing and the experimental Playgrounds package,
neither of which uses (or needs to use) a move-only type here. Earlier,
`ExitTest` conformed to `DiscoverableAsTestContent`, but this was
changed and is no longer necessary.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.