Infer the types of function/closure arguments when captured by an exit test. (#1130)
This PR adds the ability to infer the type of a parameter of a function
or closure that encloses an exit test. For example, `x` here:

```swift
func f(x: Int) async {
  await #expect(processExitsWith: .failure) { [x] in
    ...
  }
}
```

This inference still fails if a parameter is shadowed by a variable with
an incompatible type; we still need something like `decltype()` to solve
for such cases. We emit a custom diagnostic of the form "🛑 Type of
captured value 'x' is ambiguous" if the inferred type of the captured
value doesn't match what the compiler thinks it is (see #1146). Still,
being able to capture `@Test` function arguments with minimal ceremony
is helpful:

```swift
@Test(arguments: 0 ..< 100) func f(i: Int) async {
  await #expect(exitsWith: .failure) { [i] in
    ...
  }
}
```

Also type inference for literals because "why not?"

> [!NOTE]
> Exit test value capture remains an experimental feature.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.