Improve type inference for exit test value captures.
This PR refactors the (new) type inference logic for exit test capture lists to
use a syntax visitor, which allows for the types of more complex expressions to
be inferred.

For example, previously the type of this capture would not be inferred:

```swift
[x = try await f() as Int]
```

Even though the type (`Int`) is clearly present, because the `AsExprSyntax` is
nested in an `AwaitExprSyntax` and then a `TryExprSyntax`.