Improve type inference for exit test value captures. (#1163)
This PR refactors the (new) type inference logic for exit test capture
lists to use a syntax visitor, which allows for the types of more
complex expressions to be inferred.

For example, previously the type of this capture would not be inferred:

```swift
[x = try await f() as Int]
```

Even though the type (`Int`) is clearly present, because the
`AsExprSyntax` is nested in an `AwaitExprSyntax` and then a
`TryExprSyntax`.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.