Diagnose when we incorrectly infer the type of a capture list item in an exit test. (#1152)
Follow-up to #1130, split out for clarity.

This PR adds a custom diagnostic at compile time if we incorrectly infer
the type of a captured function argument or `self` in an exit test. For
example:

```swift
func f(_ x: Int) async {
  let x = String(x) // local type of 'x' is String, not Int
  await #expect(processExitsWith: ...) { [x] in
    ...
  }
}
```

This improves our feedback to the developer when we encounter a pattern
like that. The developer will now see:

> 🛑 Type of captured value 'x' is ambiguous

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.