Work around a compiler regression affecting exit test value capturing.
This PR works around a compiler regression as of Swift acbdfef4f4d71b1 that
causes our helper macro `#__capturedValue()` to be incorrectly expanded (wrong
overload is selected.) The workaround is to disable our use of this macro, which
does not affect correct functioning of our code but which does degrade the
quality of our diagnostics.

Works around rdar://154221449.