FreeBSD/OpenBSD: Gate GNU-only API
FreeBSD and OpenBSD currently use the Glibc modulemap to import the C
runtimes so `canImport(Glibc)` returns true. These systems are not
actually glibc though and do not have the `gnu_get_libc_version`
extension, resulting in build failures.

Gating the use of the API on more than glibc to allow these platforms to
compile.