Adopt `SuspendingClock.systemEpoch`. (#1202)
This PR adopts the new `systemEpoch` API added with
[SE-0473](https://github.com/swiftlang/swift-evolution/blob/main/proposals/0473-clock-epochs.md).

This API is back-deployed, so we don't need availability annotations on
Darwin. It is not available with the 6.1 toolchain though, so the old
`unsafeBitCast()` calls remain if building from source with a 6.1
toolchain.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.