Clarify the outcome of applying a timeLimit trait to a suite. (#1225)
Make it clearer that the time limit applies to each test or case in the
suite, not to the overall runtime of the test suite.

### Motivation:

I received feedback that a developer couldn't work out whether
`.timeLimit(_:)` on a suite limits each test, or the whole suite.

### Modifications:

- Add documentation that explains that the time limit applies to each
test or case in a parameterized test, when you apply it to a suite.
- Remove some passive-voice sentences in the same documents so that it's
clearer what a developer does, and what the testing library does for
them.

### Checklist:

- [X] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [X] If public symbols are renamed or modified, DocC references should
be updated.