Don't use `_getSuperclass()` from the Swift runtime. (#1237)
To determine if a class is a subclass of another, we can use `T.self is
U.Type` in an implicitly-opened existential context. Prior to the Swift
6 language mode, type existentials weren't openable, so we had to
recursively call a runtime function to check this. That's no longer a
concern for us.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.