Introduce a new `ABI.VersionNumber` type.
This PR introduces and plumbs through an internal `ABI.VersionNumber` type that
represents the JSON event stream version that should be used during a test run.
This type supports integers (such as the current version 0) and semantic
versions (e.g. 6.3 or 6.3.0). Care is taken to ensure backwards compatibility
for ABI versions -1 and 0 which do not accept a string in this position.

> [!WARNING]
> For technical reasons involving `__CommandLineArguments_v0`, this type is
> exposed publicly as `__ABIVersionNumber`. This typealias may be removed in a
> future update.

The new type necessarily conforms to `Comparable` and `Codable`. It also
conforms to `ExpressibleByIntegerLiteral` for compatibility with older versions
of Xcode that set the `__CommandLineArguments_v0.eventStreamVersion` property to
an integer literal value.

For the moment, these are all implementation details that have no impact on our
supported interface as there is no newer JSON schema than 0. We intend to make
changes to the schema in the next Swift release (assumed to be 6.3) that will
require a Swift Evolution review.