Introduce a new `ABI.VersionNumber` type. (#1239)
This PR introduces and plumbs through an internal `ABI.VersionNumber`
type that represents the JSON event stream version that should be used
during a test run. This type supports integers (such as the current
version 0) and semantic versions (e.g. 6.3 or 6.3.0). Care is taken to
ensure backwards compatibility for ABI versions -1 and 0 which do not
accept a string in this position.

The new type necessarily conforms to `Comparable` and `Codable`. I've
adjusted `__CommandLineArguments_v0` to support this new type while
maintaining backwards-compatibility with Xcode 16.

For the moment, these are all implementation details that have no impact
on our supported interface as there is no newer JSON schema than 0. We
intend to make changes to the schema in the next Swift release (assumed
to be 6.3) that will require a Swift Evolution review.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.