Allow attaching an `IWICBitmapSource` instance directly.
This PR adjusts the experimental WIC bitmap attachment support I recently added
so that a test author can attach an instance of `IWICBitmapSource` (which is the
parent type of the already-supported `IWICBitmap`.) Protocols and conformances
are adjusted to match.

I could have left the protocols as-is, but then you'd have to convert an
`IWICBitmapSource` to an `IWICBitmap` (by allocating a new COM object) then cast
it back to an `IWICBitmapSource` (by calling `QueryInterface()` and juggling the
refcount) in order to actually attach it.