Allow attaching an `IWICBitmapSource` instance directly. (#1266)
This PR adjusts the experimental WIC bitmap attachment support I
recently added so that a test author can attach an instance of
`IWICBitmapSource` (which is the parent type of the already-supported
`IWICBitmap`.) Protocols and conformances are adjusted to match.

This PR then explicitly adds conformances to every COM class in WIC that
subclasses `IWICBitmapSource` since we can't see COM class inheritance
in Swift (yet?)

I could have left the protocols as-is, but then you'd have to convert an
`IWICBitmapSource` to an `IWICBitmap` (by allocating a new COM object)
then cast it back to an `IWICBitmapSource` (by calling
`QueryInterface()` and juggling the refcount) in order to actually
attach it.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.