Reduce duplicated code between Darwin's and Windows' image attachments.
This PR cleans up the implementations of image attachments on Darwin and Windows
so that we have less code duplication between the two. `_AttachableImageWrapper`
is partially lowered to the main library (excepting the platform-specific bits)
and `ImageAttachmentError` is lowered in its entirety.

There are some adjustments to the (internal/package) interface of
`_AttachableImageWrapper` to accomodate it not being able to specify conformance
to `AttachableAsCGImage` or `AttachableAsIWICBitmapSource`. Namely, whatever
code initializes an instance of it is responsible for copying `image` and
providing a deinitializer function as needed.