Ensure attachments created in exit tests are forwarded to the parent.
This PR ensures that when an exit test records an attachment, it is forwarded to
the parent process.

It introduces a local/private dependency on `Data` in `EncodedAttachment` so
that we can use its base64 encoding/decoding and its ability to map files from
disk. If Foundation is not available, it falls back to encoding/decoding
`[UInt8]` and reading files into memory with `fread()` (the old-fashioned way).

Resolves rdar://149242118.