Add experimental SPI to cancel a running test.
This PR introduces `Test.cancel()` and `Test.Case.cancel()` which cancel the
current test/suite and the current test case, respectively. For example:

```swift
@Test(arguments: [Food.burger, .fries, .iceCream])
func `Food truck is well-stocked`(_ food: Food) throws {
  if food == .iceCream && Season.current == .winter {
    try Test.Case.cancel("It's too cold for ice cream.")
  }
  // ...
}
```

These functions work by cancelling the child task associated with the current
test or test case, then throwing an error to end local execution early. Compare
`XCTSkip()` which, in Swift, is just a thrown error that the XCTest harness
special-cases, or `XCTSkip()` in Objective-C which actually throws an exception
to force the caller to exit early.