Rename `clsid` to `encoderCLSID` and add label to corresponding `init()`. (#1295)
This PR adjusts the (experimental) extensions to `AttachableImageFormat`
used for Windows support so that the `CLSID` value we use is explicitly
defined in the API (not just documentation) as the _encoder_ CLSID. The
Windows SDK also specifies _decoder_ CLSIDs and _container GUIDs_ which
could be mistakenly passed here but which won't work correctly.

Also some tweaks to documentation on the Apple side to match changes on
the Windows side.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.