Move the package version to version.txt.
This change moves the definition of Swift Testing's version to a new file,
version.txt. This file doesn't represent a standard, _per se_, but it is common
for open-source projects to list their version info in such a file. And by
putting our version info in this file, we can share it between package builds
and CMake builds.