Avoid eagerly serializing move-only attachments.
This PR adjusts the interface of `Attachment` so that it always conforms to
`Copyable` even if the value it wraps does not. This allows us to avoid eagerly
serializing attachments that are not `Copyable` but _are_ `Sendable`. We must
still eagerly serialize non-sendable attachments.

This change will allow us to support "attachment lifetime" (API to be designed,
see XCTest's API [here](https://developer.apple.com/documentation/xctest/xctattachment/lifetime-swift.enum))
more efficiently and for more attachable values. If you're paying the cost of
serialization for attachments you don't end up keeping, that's a waste of\
resources.