Work around lack of embed in OpenBSD's clang.
For various reasons, the Swift toolchain for OpenBSD relies on using the
platform's native clang, which is 16. clang 19 is the most recent version
that will not emit an error with the new __has_embed features in C23.

Since swift-testing is experimentally supported by OpenBSD and thus to
make swift-testing build again on the platform, work around the issue
with a platform-specific command-line specified macro override in swiftpm
and in cmake.

Furthermore, we can use cmake trickery to subsitute the version file
contents instead of using embed. This may not be possible to do with
swiftpm, but I don't know for sure.