Merge more of the image attachments codebase duplicated between Darwin and Windows.
This PR eliminates a bunch of duplicated code between Darwin and Windows that's
used to support image attachments. We're able to do this because
`_AttachableImageWrapper` is now a class and can be included in conforms-to
generic constraints, so we can write
`where AttachableValue: _AttachableImageWrapper<T> & AttachableWrapper`. It was
previously a structure and structures can't be used in generic constraints this
way.

There's a new public, underscored protocol introduced here, `_AttachableAsImage`,
that serves as a "base" protocol for `AttachableAsCGImage` and
`AttachableAsIWICBitmapSource`. I intend to promote this protocol to API, but it
will need a Swift Evolution proposal first. I'm going to include it in a future
"image attachments refinement" proposal that will cover a few other things.