Change the semantics of `Test.cancel()`.
This PR changes the semantics of `Test.cancel()` such that it only cancels the
current test case if the current test is parameterized. This PR also removes
`Test.Case.cancel()`, but we may opt to add `Test.Case.cancelAll()` or some such
in the future to re-add an interface that cancels an entire parameterized test.