Promote `SourceLocation._filePath` to proper, non-underscored SPI.
This PR introduces an `@_spi var filePath` property on `SourceLocation` with the
intent of replacing the existing underscored/unsupported `_filePath` property.

Because Xcode 16 through 26 directly encodes and decodes instances of
`SourceLocation`, we can't just stop encoding/decoding the `_filePath` key and
need to keep it there for at least a Swift release or two, until Apple's fork of
Swift Testing has been updated to include the new encoding key. At that point we
can remove the old one.